+++ lib/flutter_carbon.dart
@@ -38,6 +38,7 @@ export 'src/widgets/carbon_tile.dart';
 export 'src/widgets/carbon_multi_select.dart';
 export 'src/widgets/carbon_contained_list.dart';
 export 'src/widgets/carbon_floating_menu.dart';
+export 'src/widgets/carbon_tabs.dart';
 
 // Component Themes
 export 'src/theme/component_themes/ai_theme_data.dart';